Updated DTO and  verified 4.3  in POSTMAN

+++ src/main/java/org/cognizant/disastermanagement/dao/EmergencyRepository.java
@@ -3,5 +3,5 @@ package org.cognizant.disastermanagement.dao;
 import org.cognizant.disastermanagement.entity.EmergencyReport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmergencyRepository extends JpaRepository<EmergencyReport, Long> {
+public interface EmergencyRepository extends JpaRepository<EmergencyReport, Integer> {
 }
