+++ src/main/java/org/cognizant/disastermanagement/dao/IncidentRepository.java
@@ -3,5 +3,5 @@ package org.cognizant.disastermanagement.dao;
 import org.cognizant.disastermanagement.entity.Incident;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IncidentRepository extends JpaRepository<Incident, Long> {
+public interface IncidentRepository extends JpaRepository<Incident, Integer> {
 }
